backend/requirements-dev.txt
-r requirements.txt
pytest==8.1.1
pytest-cov==5.0.0
python-socketio[client]==5.8.0
